@@ -1,4 +1,4 @@
-@model ErrorViewModel
+﻿@model ErrorViewModel
 @{
     ViewData["Title"] = "Error";
 }
